requirements.txt
fastapi
uvicorn
pyodbc
python-dotenv
qrcode
cryptography
pycryptodome
twilio
email-validator
razorpay